// src/domicilios-ciudadano/dto/create-domicilios-ciudadano.dto.ts
export class CreateDomiciliosCiudadanoDto {

    ciudadano_id: number;
        
    pais_id: string;
    
    provincia_id: string;
    
    departamento_id: number;
    
    municipio_id: number;
   
    ciudad: string;
   
    barrio: string;
   
    direccion: string;
    
    numero_dom: number;

    detalle_motivo: string;
      
    fecha_cambio: Date;
    
    organismo_id: number;
    
    usuario_id: number;
    
}